tests/requirements.txt
selenium
pytest
pytest-xdist
pytest-cov
httpx
webdriver-manager 
requests
chromedriver-autoinstallerpip 